+++ examples/applications/tropic01_util/firmware/Src/main.c
@@ -1,0 +1,491 @@
+/* USER CODE BEGIN Header */
+/**
+ * @file main.c
+ * @brief Main source file for the tropic01_util application on TROPIC01 USB DevKit.
+ * @copyright Copyright (c) 2020-2026 Tropic Square s.r.o.
+ *
+ * @license For the license see LICENSE.md in the root directory of this source tree.
+ */
+
+// This file was generated using Stm32CubeMX and modified by Tropic Square.
+
+/**
+ * Copyright (c) 2026 STMicroelectronics.
+ * All rights reserved.
+ *
+ * This software is licensed under terms that can be found in the LICENSE file
+ * in the root directory of this software component.
+ * If no LICENSE file comes with this software, it is provided AS-IS.
+ */
+/* USER CODE END Header */
+/* Includes ------------------------------------------------------------------*/
+#include "main.h"
+
+#include <stdbool.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+
+#include "frame_protocol.h"
+#include "libtropic.h"
+#include "libtropic_logging.h"
+#include "libtropic_mbedtls_v4.h"
+#include "libtropic_port_stm32u5xx.h"
+#include "psa/crypto.h"
+#include "tusb.h"
+
+/* Private includes ----------------------------------------------------------*/
+/* USER CODE BEGIN Includes */
+
+/* USER CODE END Includes */
+
+/* Private typedef -----------------------------------------------------------*/
+/* USER CODE BEGIN PTD */
+
+/* USER CODE END PTD */
+
+/* Private define ------------------------------------------------------------*/
+/* USER CODE BEGIN PD */
+
+#define CHECK_LT_RET           \
+    {                          \
+        if (LT_OK != lt_ret) { \
+            Error_Handler();   \
+        }                      \
+    }
+
+/* USER CODE END PD */
+
+/* Private macro -------------------------------------------------------------*/
+/* USER CODE BEGIN PM */
+
+/* USER CODE END PM */
+
+/* Private variables ---------------------------------------------------------*/
+
+CRC_HandleTypeDef hcrc;
+
+DCACHE_HandleTypeDef hdcache1;
+
+RNG_HandleTypeDef hrng;
+
+HASH_HandleTypeDef hhash;
+
+SPI_HandleTypeDef *hspi1 = NULL;
+
+PCD_HandleTypeDef hpcd_USB_DRD_FS;
+
+/* USER CODE BEGIN PV */
+lt_handle_t lt_handle = {0};
+
+/* USER CODE END PV */
+
+/* Private function prototypes -----------------------------------------------*/
+void SystemClock_Config(void);
+static void MX_GPIO_Init(void);
+static void MX_CRC_Init(void);
+static void MX_DCACHE1_Init(void);
+static void MX_ICACHE_Init(void);
+static void MX_RNG_Init(void);
+static void MX_USB_DRD_FS_PCD_Init(void);
+
+#if defined(__ICCARM__)
+/* New definition from EWARM V9, compatible with EWARM8 */
+int iar_fputc(int ch);
+#define PUTCHAR_PROTOTYPE int iar_fputc(int ch)
+#elif defined(__CC_ARM) || defined(__ARMCC_VERSION)
+/* ARM Compiler 5/6*/
+#define PUTCHAR_PROTOTYPE int fputc(int ch, FILE *f)
+#elif defined(__GNUC__)
+#define PUTCHAR_PROTOTYPE int __io_putchar(int ch)
+#endif /* __ICCARM__ */
+
+/* USER CODE BEGIN PFP */
+
+/* USER CODE END PFP */
+
+/* Private user code ---------------------------------------------------------*/
+/* USER CODE BEGIN 0 */
+
+/* USER CODE END 0 */
+
+/**
+ * @brief  The application entry point.
+ * @retval int
+ */
+int main(void)
+{
+    /* USER CODE BEGIN 1 */
+
+    /* USER CODE END 1 */
+
+    /* MCU Configuration--------------------------------------------------------*/
+
+    /* Reset of all peripherals, Initializes the Flash interface and the Systick. */
+    HAL_Init();
+
+    /* USER CODE BEGIN Init */
+
+    /* USER CODE END Init */
+
+    /* Configure the system clock */
+    SystemClock_Config();
+
+    /* USER CODE BEGIN SysInit */
+
+    /* USER CODE END SysInit */
+
+    /* Initialize all configured peripherals */
+    // Note: SPI is initialized by Libtropic.
+    MX_GPIO_Init();
+    MX_CRC_Init();
+    MX_DCACHE1_Init();
+    MX_ICACHE_Init();
+    MX_RNG_Init();
+    MX_USB_DRD_FS_PCD_Init();
+    /* USER CODE BEGIN 2 */
+
+    // Initialize USB.
+    if (!tusb_init()) {
+        Error_Handler();
+    }
+
+    // Power up TROPIC01.
+    HAL_GPIO_WritePin(TR01_PWR_GPIO_Port, TR01_PWR_Pin, GPIO_PIN_SET);
+
+    // Setup Libtropic.
+    lt_dev_stm32u5xx_t lt_device = {0};
+    lt_ctx_mbedtls_v4_t lt_crypto_ctx = {0};
+    lt_ret_t lt_ret;
+
+    lt_device.spi_instance = SPI1;
+    lt_device.baudrate_prescaler = SPI_BAUDRATEPRESCALER_32;
+    lt_device.spi_cs_gpio_bank = TR01_CS_GPIO_Port;
+    lt_device.spi_cs_gpio_pin = TR01_CS_Pin;
+    lt_device.rng_handle = &hrng;
+#ifdef LT_USE_INT_PIN
+    lt_device.int_gpio_bank = TR01_GPO_GPIO_Port;
+    lt_device.int_gpio_pin = TR01_GPO_Pin;
+#endif
+
+    lt_handle.l2.device = &lt_device;
+    lt_handle.l3.crypto_ctx = &lt_crypto_ctx;
+
+    // Initialize MbedTLS.
+    psa_status_t status = psa_crypto_init();
+    if (status != PSA_SUCCESS) {
+        Error_Handler();
+    }
+
+    // Initialize Libtropic.
+    lt_ret = lt_init(&lt_handle);
+    CHECK_LT_RET;
+
+    lt_ret = lt_verify_chip_and_start_secure_session(&lt_handle, lt_sh0priv_prod0, lt_sh0pub_prod0,
+                                                     TR01_PAIRING_KEY_SLOT_INDEX_0);
+    CHECK_LT_RET;
+
+    // Save pointer to the SPI handle so we can use SPI with raw commands.
+    hspi1 = &lt_device.spi_handle;
+
+    // Turn on the LED to indicate that the initialization succeeded.
+    HAL_GPIO_WritePin(APP_LED_GPIO_Port, APP_LED_Pin, GPIO_PIN_SET);
+
+    usb_devkit_main_loop();
+    /* USER CODE END 2 */
+}
+
+/**
+ * @brief System Clock Configuration
+ * @retval None
+ */
+void SystemClock_Config(void)
+{
+    RCC_OscInitTypeDef RCC_OscInitStruct = {0};
+    RCC_ClkInitTypeDef RCC_ClkInitStruct = {0};
+
+    /** Configure the main internal regulator output voltage
+     */
+    if (HAL_PWREx_ControlVoltageScaling(PWR_REGULATOR_VOLTAGE_SCALE1) != HAL_OK) {
+        Error_Handler();
+    }
+
+    /** Initializes the CPU, AHB and APB buses clocks
+     */
+    RCC_OscInitStruct.OscillatorType = RCC_OSCILLATORTYPE_HSI48 | RCC_OSCILLATORTYPE_HSE;
+    RCC_OscInitStruct.HSEState = RCC_HSE_ON;
+    RCC_OscInitStruct.HSI48State = RCC_HSI48_ON;
+    RCC_OscInitStruct.PLL.PLLState = RCC_PLL_ON;
+    RCC_OscInitStruct.PLL.PLLSource = RCC_PLLSOURCE_HSE;
+    RCC_OscInitStruct.PLL.PLLMBOOST = RCC_PLLMBOOST_DIV1;
+    RCC_OscInitStruct.PLL.PLLM = 1;
+    RCC_OscInitStruct.PLL.PLLN = 20;
+    RCC_OscInitStruct.PLL.PLLP = 2;
+    RCC_OscInitStruct.PLL.PLLQ = 2;
+    RCC_OscInitStruct.PLL.PLLR = 1;
+    RCC_OscInitStruct.PLL.PLLRGE = RCC_PLLVCIRANGE_1;
+    RCC_OscInitStruct.PLL.PLLFRACN = 0;
+    if (HAL_RCC_OscConfig(&RCC_OscInitStruct) != HAL_OK) {
+        Error_Handler();
+    }
+
+    /** Initializes the CPU, AHB and APB buses clocks
+     */
+    RCC_ClkInitStruct.ClockType = RCC_CLOCKTYPE_HCLK | RCC_CLOCKTYPE_SYSCLK | RCC_CLOCKTYPE_PCLK1 |
+                                  RCC_CLOCKTYPE_PCLK2 | RCC_CLOCKTYPE_PCLK3;
+    RCC_ClkInitStruct.SYSCLKSource = RCC_SYSCLKSOURCE_PLLCLK;
+    RCC_ClkInitStruct.AHBCLKDivider = RCC_SYSCLK_DIV1;
+    RCC_ClkInitStruct.APB1CLKDivider = RCC_HCLK_DIV1;
+    RCC_ClkInitStruct.APB2CLKDivider = RCC_HCLK_DIV1;
+    RCC_ClkInitStruct.APB3CLKDivider = RCC_HCLK_DIV1;
+
+    if (HAL_RCC_ClockConfig(&RCC_ClkInitStruct, FLASH_LATENCY_4) != HAL_OK) {
+        Error_Handler();
+    }
+}
+
+/**
+ * @brief CRC Initialization Function
+ * @param None
+ * @retval None
+ */
+static void MX_CRC_Init(void)
+{
+    /* USER CODE BEGIN CRC_Init 0 */
+
+    /* USER CODE END CRC_Init 0 */
+
+    /* USER CODE BEGIN CRC_Init 1 */
+
+    /* USER CODE END CRC_Init 1 */
+    hcrc.Instance = CRC;
+    hcrc.Init.DefaultPolynomialUse = DEFAULT_POLYNOMIAL_DISABLE;
+    hcrc.Init.DefaultInitValueUse = DEFAULT_INIT_VALUE_DISABLE;
+    hcrc.Init.GeneratingPolynomial = 32773;
+    hcrc.Init.CRCLength = CRC_POLYLENGTH_16B;
+    hcrc.Init.InitValue = 0;
+    hcrc.Init.InputDataInversionMode = CRC_INPUTDATA_INVERSION_NONE;
+    hcrc.Init.OutputDataInversionMode = CRC_OUTPUTDATA_INVERSION_DISABLE;
+    hcrc.InputDataFormat = CRC_INPUTDATA_FORMAT_BYTES;
+    if (HAL_CRC_Init(&hcrc) != HAL_OK) {
+        Error_Handler();
+    }
+    /* USER CODE BEGIN CRC_Init 2 */
+
+    /* USER CODE END CRC_Init 2 */
+}
+
+/**
+ * @brief DCACHE1 Initialization Function
+ * @param None
+ * @retval None
+ */
+static void MX_DCACHE1_Init(void)
+{
+    /* USER CODE BEGIN DCACHE1_Init 0 */
+
+    /* USER CODE END DCACHE1_Init 0 */
+
+    /* USER CODE BEGIN DCACHE1_Init 1 */
+
+    /* USER CODE END DCACHE1_Init 1 */
+    hdcache1.Instance = DCACHE1;
+    hdcache1.Init.ReadBurstType = DCACHE_READ_BURST_WRAP;
+    if (HAL_DCACHE_Init(&hdcache1) != HAL_OK) {
+        Error_Handler();
+    }
+    /* USER CODE BEGIN DCACHE1_Init 2 */
+
+    /* USER CODE END DCACHE1_Init 2 */
+}
+
+/**
+ * @brief ICACHE Initialization Function
+ * @param None
+ * @retval None
+ */
+static void MX_ICACHE_Init(void)
+{
+    /* USER CODE BEGIN ICACHE_Init 0 */
+
+    /* USER CODE END ICACHE_Init 0 */
+
+    /* USER CODE BEGIN ICACHE_Init 1 */
+
+    /* USER CODE END ICACHE_Init 1 */
+
+    /** Enable instruction cache (default 2-ways set associative cache)
+     */
+    if (HAL_ICACHE_Enable() != HAL_OK) {
+        Error_Handler();
+    }
+    /* USER CODE BEGIN ICACHE_Init 2 */
+
+    /* USER CODE END ICACHE_Init 2 */
+}
+
+/**
+ * @brief RNG Initialization Function
+ * @param None
+ * @retval None
+ */
+static void MX_RNG_Init(void)
+{
+    /* USER CODE BEGIN RNG_Init 0 */
+
+    /* USER CODE END RNG_Init 0 */
+
+    /* USER CODE BEGIN RNG_Init 1 */
+
+    /* USER CODE END RNG_Init 1 */
+    hrng.Instance = RNG;
+    hrng.Init.ClockErrorDetection = RNG_CED_ENABLE;
+    if (HAL_RNG_Init(&hrng) != HAL_OK) {
+        Error_Handler();
+    }
+    /* USER CODE BEGIN RNG_Init 2 */
+
+    /* USER CODE END RNG_Init 2 */
+}
+
+/**
+ * @brief USB_DRD_FS Initialization Function
+ * @param None
+ * @retval None
+ */
+static void MX_USB_DRD_FS_PCD_Init(void)
+{
+    /* USER CODE BEGIN USB_DRD_FS_Init 0 */
+
+    /* USER CODE END USB_DRD_FS_Init 0 */
+
+    /* USER CODE BEGIN USB_DRD_FS_Init 1 */
+
+    /* USER CODE END USB_DRD_FS_Init 1 */
+    hpcd_USB_DRD_FS.Instance = USB_DRD_FS;
+    hpcd_USB_DRD_FS.Init.dev_endpoints = 8;
+    hpcd_USB_DRD_FS.Init.speed = PCD_SPEED_FULL;
+    hpcd_USB_DRD_FS.Init.phy_itface = PCD_PHY_EMBEDDED;
+    hpcd_USB_DRD_FS.Init.Sof_enable = DISABLE;
+    hpcd_USB_DRD_FS.Init.low_power_enable = DISABLE;
+    hpcd_USB_DRD_FS.Init.lpm_enable = DISABLE;
+    hpcd_USB_DRD_FS.Init.battery_charging_enable = DISABLE;
+    hpcd_USB_DRD_FS.Init.vbus_sensing_enable = DISABLE;
+    hpcd_USB_DRD_FS.Init.bulk_doublebuffer_enable = DISABLE;
+    hpcd_USB_DRD_FS.Init.iso_singlebuffer_enable = DISABLE;
+    if (HAL_PCD_Init(&hpcd_USB_DRD_FS) != HAL_OK) {
+        Error_Handler();
+    }
+    /* USER CODE BEGIN USB_DRD_FS_Init 2 */
+
+    // Map the Packet Memory Area (PMA) for the CDC ACM Endpoints
+    HAL_PCDEx_PMAConfig(&hpcd_USB_DRD_FS, 0x00, PCD_SNG_BUF, 0x14);   // 0x00 = EP0 OUT
+    HAL_PCDEx_PMAConfig(&hpcd_USB_DRD_FS, 0x80, PCD_SNG_BUF, 0x54);   // 0x80 = EP0 IN
+    HAL_PCDEx_PMAConfig(&hpcd_USB_DRD_FS, 0x81, PCD_SNG_BUF, 0x94);   // 0x81 = EP1 IN (Data)
+    HAL_PCDEx_PMAConfig(&hpcd_USB_DRD_FS, 0x01, PCD_SNG_BUF, 0xD4);   // 0x01 = EP1 OUT (Data)
+    HAL_PCDEx_PMAConfig(&hpcd_USB_DRD_FS, 0x82, PCD_SNG_BUF, 0x114);  // 0x82 = EP2 IN (Cmd)
+
+    /* USER CODE END USB_DRD_FS_Init 2 */
+}
+
+/**
+ * @brief GPIO Initialization Function
+ * @param None
+ * @retval None
+ */
+static void MX_GPIO_Init(void)
+{
+    GPIO_InitTypeDef GPIO_InitStruct = {0};
+    /* USER CODE BEGIN MX_GPIO_Init_1 */
+
+    /* USER CODE END MX_GPIO_Init_1 */
+
+    /* GPIO Ports Clock Enable */
+    __HAL_RCC_GPIOH_CLK_ENABLE();
+    __HAL_RCC_GPIOA_CLK_ENABLE();
+    __HAL_RCC_GPIOB_CLK_ENABLE();
+
+    /*Configure GPIO pin Output Level */
+    // We set the USB_DP_Pin here to 0 to force the USB host to re-enumerate when the device resets.
+    // This is needed after entering the DFU Mode, so the user doesn't have to replug the device after
+    // a firmware update.
+    HAL_GPIO_WritePin(GPIOA, TR01_PWR_Pin | APP_LED_Pin | USB_DP_Pin, GPIO_PIN_RESET);
+    HAL_GPIO_WritePin(TR01_CS_GPIO_Port, TR01_CS_Pin, GPIO_PIN_SET);
+
+    /*Configure GPIO pins : TR01_PWR_Pin TR01_CS_Pin APP_LED_Pin */
+    GPIO_InitStruct.Pin = TR01_PWR_Pin | TR01_CS_Pin | APP_LED_Pin | USB_DP_Pin;
+    GPIO_InitStruct.Mode = GPIO_MODE_OUTPUT_PP;
+    GPIO_InitStruct.Pull = GPIO_NOPULL;
+    GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_LOW;
+    HAL_GPIO_Init(GPIOA, &GPIO_InitStruct);
+
+    /*Configure GPIO pin : TR01_GPO_Pin */
+    GPIO_InitStruct.Pin = TR01_GPO_Pin;
+    GPIO_InitStruct.Mode = GPIO_MODE_INPUT;
+    GPIO_InitStruct.Pull = GPIO_NOPULL;
+    HAL_GPIO_Init(TR01_GPO_GPIO_Port, &GPIO_InitStruct);
+
+    /* USER CODE BEGIN MX_GPIO_Init_2 */
+
+    /* USER CODE END MX_GPIO_Init_2 */
+}
+
+/* USER CODE BEGIN 4 */
+
+#if USB_DEVKIT_DEBUG_CDC_ITF
+/**
+ * @brief  Retargets the C library printf function to the USB.
+ * @param  None
+ * @retval None
+ */
+PUTCHAR_PROTOTYPE
+{
+    /*  Translates LF to CFLF, as this is what most serial monitors expect
+        by default
+    */
+    if (ch == '\n') {
+        tud_cdc_n_write(DEBUG_CDC_ITF, "\r\n", 2);
+    }
+    else {
+        tud_cdc_n_write(DEBUG_CDC_ITF, &ch, 1);
+    }
+
+    tud_cdc_n_write_flush(DEBUG_CDC_ITF);
+    return ch;
+}
+#endif
+
+/* USER CODE END 4 */
+
+/**
+ * @brief  This function is executed in case of error occurrence.
+ * @retval None
+ */
+void Error_Handler(void)
+{
+    /* USER CODE BEGIN Error_Handler_Debug */
+    /* User can add his own implementation to report the HAL error return state */
+    while (1) {
+        HAL_GPIO_WritePin(APP_LED_GPIO_Port, APP_LED_Pin, GPIO_PIN_SET);
+        HAL_Delay(500);
+        HAL_GPIO_WritePin(APP_LED_GPIO_Port, APP_LED_Pin, GPIO_PIN_RESET);
+        HAL_Delay(500);
+    }
+    /* USER CODE END Error_Handler_Debug */
+}
+#ifdef USE_FULL_ASSERT
+/**
+ * @brief  Reports the name of the source file and the source line number
+ *         where the assert_param error has occurred.
+ * @param  file: pointer to the source file name
+ * @param  line: assert_param error line source number
+ * @retval None
+ */
+void assert_failed(uint8_t *file, uint32_t line)
+{
+    /* USER CODE BEGIN 6 */
+    /* User can add his own implementation to report the file name and line number,
+       ex: printf("Wrong parameters value: file %s on line %d\r\n", file, line) */
+    /* USER CODE END 6 */
+}
+#endif /* USE_FULL_ASSERT */